pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hmdl"
version = "0.1.0"
description = "Observability SDK for MCP (Model Context Protocol) servers - Heimdall Platform"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Heimdall Team", email = "team@heimdall.dev" }
]
keywords = [
    "observability",
    "mcp",
    "model-context-protocol",
    "opentelemetry",
    "tracing",
    "monitoring",
    "llm",
    "ai"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.9"
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-semantic-conventions>=0.41b0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/heimdall-observability/heimdall"
Documentation = "https://docs.heimdall.dev"
Repository = "https://github.com/heimdall-observability/heimdall"
Issues = "https://github.com/heimdall-observability/heimdall/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/hmdl"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

